Completes separation of Contexts
All dispatch calls and state values separated by Global/'Main' concerns.

Main context is where working tree object and display package is stored.

Global context holds user data, ace editor data, etc..
