+++ src/components/ace-window/ace-window.component.tsx
@@ -13,8 +13,8 @@
 ******************************************************************************/
 
 import { useContext, useEffect, useRef } from 'react'
-import { MainContext } from '../../context/main/MainState'
-
+// import { MainContext } from '../../context/main/MainState'
+import { GlobalContext } from '../../context/global/GlobalState'
 import { 
   /* Assets */
   /* Database */
@@ -26,10 +26,14 @@ import {
 import './ace-window.styles.scss'
 
 const AceWindow = (props: any): JSX.Element => {
+  // const {
+  //   state: {  },
+  // } = useContext(MainContext)
+
   const {
-    state: { aceObj, editorPacket },
-    dispatch,
-  } = useContext(MainContext)
+    state: { editorPacket, aceObj },
+    globalDispatch,
+  } = useContext(GlobalContext)
 
   const currentEditor = useRef(null)
 
@@ -60,7 +64,7 @@ const AceWindow = (props: any): JSX.Element => {
     props.id,
     Ace,
     props.codeContent,
-    dispatch,
+    globalDispatch,
     props.language,
   ])
