+++ src/context/global/GlobalReducer.ts
@@ -24,6 +24,7 @@ import {
 export const GlobalReducer = (state: any, action: any) => {
 
   switch (action.type) {
+
     case 'SET_CURRENT_USER_TO_STATE': {
       // console.log(`Trace: SET_CURRENT_USER_TO_STATE()`)
       let data = action.payload.userObj
@@ -94,7 +95,6 @@ export const GlobalReducer = (state: any, action: any) => {
         currentDropChain: [...action.payload.chain],
         currentDropPaneChain: null,
       }
-      // console.log(action.payload.currentDropId)
       return {
         ...state,
         globalDragData: globalDragData,
@@ -107,7 +107,6 @@ export const GlobalReducer = (state: any, action: any) => {
         ...state.globalDragData,
         currentDraggingId: action.payload.currentDraggingId,
       }
-      // console.log(action.payload.currentDropId)
       return {
         ...state,
         globalDragData: globalDragData,
@@ -123,7 +122,6 @@ export const GlobalReducer = (state: any, action: any) => {
         currentDropPaneChain: [...action.payload.chain],
         currentDropChain: null,
       }
-      // console.log(action.payload.currentDropId)
       return {
         ...state,
         globalDragData: globalDragData,
@@ -149,7 +147,6 @@ export const GlobalReducer = (state: any, action: any) => {
     }
 
     case 'CREATE_PRIMARY': {
-      // global
       // console.log(`Trace: CREATE_PRIMARY()`)
       let primaryCategories = [...state.primaryCategories]
       primaryCategories.push(action.payload.entry)
@@ -160,11 +157,28 @@ export const GlobalReducer = (state: any, action: any) => {
     }
 
     case 'SET_PRIMARY_CATEGORIES': {
-      // global
       let primaryCategories = [...action.payload.primaryCategories]
       return {
         ...state,
         primaryCategories: primaryCategories,
+      }
+    }
+
+    case 'SEND_ENTRY_TO_EDITOR': {
+      let editorPacket = {
+        ...action.payload.editorPacket,
+      }
+      return {
+        ...state,
+        editorPacket: editorPacket,
+      }
+    }
+
+    case 'SET_ACE': {
+      // console.log(`Trace: SET_ACE()`)
+      return {
+        ...state,
+        aceObj: action.payload.aceObj,
       }
     }
