+++ src/context/global/GlobalState.tsx
@@ -44,6 +44,8 @@ interface initialStateType {
   subSubPaneEntry: any
   subPaneEntry: any
   primaryCategories: any[]
+  editorPacket: any
+  aceObj: any
 }
 
 /** INITIAL STATE DECLARATION **************************************/
@@ -66,6 +68,8 @@ const initialState = {
   subSubPaneEntry: {},
   subPaneEntry: {},
   primaryCategories: [],
+  editorPacket: {},
+  aceObj: null,
 }
 
 export const GlobalContext = createContext<{
